Update arducam_mipicamera.py
solves issue #36: variable real_width has to be of type integer in order to be processed properly in function align_up. In Python 3 the division width / 8 results in a floating point value (this is different in Python 2). The floor division assures that real_width is an integer in Python 3. The import line keeps compatibility of the code with Python 2 (should work there as well, not tested).